Fix multiple playback issues (end of file, multiple players, etc)
When Pyglet finishes playback, it will dequeue the source media. Calling
`play` again will immediately pause, because there's no source media in
the queue.

This fix addresses that issue by checking the player queue state and
requeueing the media if it expired out of the queue.

This also addresses an issue where pressing play on two spots will
allow you to play them simultaneously. Since the UI doesn't show that, I
fixed the behavior to always stop playback. I also changed the behavior
to make sure we start playback from the beginning, since the expected
behavior of a "stop" button is to stop audio and reset the transport to
the beginning of the track for future playback,
not resume playback from the original timestamp.